Use new TxnExecutor for createDelistTxnData
